Cel Shading Armageddon
A LOT of changes and experimenting. Testing ~3 different cel shaders and their effects. A lot of this needs to be culled but I haven't gotten round to sorting through all of it. Also made a start on sorting out the level meshes but this will all need to be reevaluated. Will sort out the visual style and then fully prepare everything necessary. Apologies for the massive update.