Music Manager + Meshing
Added a music + SFX manager that is super easy to add more sounds to. Temp. bgm + hit sound. The only thing with a sfx on is the Damage function, which won't now fire without audio. Is probably enough for the vertical slice but we have more options in the future. Also a fade function to swap songs.
Added more meshes to scenes and finished adding floors + walls + adjusting the camera. Boss room has camera moved further back to include more of the room. Some more stuff added I've probably forgotten to mention.